Decompiled all lara control functions
- Decompiled LaraFallen(), lara_default_col(), lara_col_kick(), GetLaraCollisionInfo(), LaraSlideSlope().
- Reduced light power of the candle, increased pitch for candle fire, reduced to 1 the sound played when candle is triggered.
- Improved light, now item is affected by room ambient.

+++ TR2Main-VS/3dsystem/phd_math.cpp
@@ -477,10 +477,10 @@ int __fastcall phd_sin(short angle) {
 	return result;
 }
 
-DWORD __fastcall phd_sqrt(DWORD n) {
-	DWORD result = 0;
-	DWORD base = 0x40000000;
-	DWORD basedResult;
+long __fastcall phd_sqrt(long n) {
+	long result = 0;
+	long base = 0x40000000;
+	long basedResult;
 
 	for (; base != 0; base >>= 2) {
 		for (; base != 0; base >>= 2) {
